# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Unila Digital Sign API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  dosen: 'dosen',
  mahasiswa: 'mahasiswa',
} as const;

/**
 * @nullable
 */
export type UserDigitalIdStatus = typeof UserDigitalIdStatus[keyof typeof UserDigitalIdStatus] | null;


export const UserDigitalIdStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  revoked: 'revoked',
  null: 'null',
} as const;

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  otpEnabled: boolean;
  /** @nullable */
  digitalIdStatus?: UserDigitalIdStatus;
  hasSignature?: boolean;
  createdAt?: string;
}

export interface AuthResponse {
  user: User;
  token: string;
  otpRequired?: boolean;
}

export interface OtpSetup {
  qrCodeUrl: string;
  secret: string;
}

export interface ChangePasswordInput {
  oldPassword: string;
  newPassword: string;
  otpCode: string;
}

export interface MessageResponse {
  message: string;
}

export interface OtpVerifyInput {
  code: string;
}

export interface OtpVerifyResponse {
  success: boolean;
  message?: string;
}

export interface OtpResetInput {
  token: string;
}

export interface CertificateInfo {
  name: string;
  issuer: string;
  validFrom: string;
  validTo: string;
  downloadUrl: string;
  instructions?: string;
}

export type DigitalIdInputRole = typeof DigitalIdInputRole[keyof typeof DigitalIdInputRole];


export const DigitalIdInputRole = {
  dosen: 'dosen',
  mahasiswa: 'mahasiswa',
} as const;

export interface DigitalIdInput {
  name: string;
  role: DigitalIdInputRole;
  email: string;
  passphrase: string;
  /** NIM (mahasiswa) or NIP (dosen) */
  nimNip?: string;
}

export type DigitalIdRequestRole = typeof DigitalIdRequestRole[keyof typeof DigitalIdRequestRole];


export const DigitalIdRequestRole = {
  dosen: 'dosen',
  mahasiswa: 'mahasiswa',
} as const;

export type DigitalIdRequestStatus = typeof DigitalIdRequestStatus[keyof typeof DigitalIdRequestStatus];


export const DigitalIdRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  revoked: 'revoked',
} as const;

export interface DigitalIdRequest {
  id: number;
  userId: number;
  userName?: string;
  name: string;
  role: DigitalIdRequestRole;
  email: string;
  /**
     * NIM (mahasiswa) or NIP (dosen)
     * @nullable
     */
  nimNip?: string | null;
  status: DigitalIdRequestStatus;
  isApproved: boolean;
  isReady: boolean;
  isSent: boolean;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface SignatureImageInput {
  /** Base64-encoded image */
  imageData: string;
}

export interface SignatureImage {
  id: number;
  userId: number;
  imageData: string;
  createdAt: string;
}

export interface DocumentInput {
  name: string;
  /** Base64-encoded file */
  fileData: string;
  fileType: string;
  pageCount?: number;
}

export type DocumentStatus = typeof DocumentStatus[keyof typeof DocumentStatus];


export const DocumentStatus = {
  uploaded: 'uploaded',
  pending_sign: 'pending_sign',
  signed: 'signed',
  rejected: 'rejected',
} as const;

export interface Document {
  id: number;
  userId: number;
  name: string;
  fileType: string;
  status: DocumentStatus;
  /** @nullable */
  pageCount?: number | null;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type UserSignRequestStatus = typeof UserSignRequestStatus[keyof typeof UserSignRequestStatus];


export const UserSignRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface UserSignRequest {
  id: number;
  documentId: number;
  /** @nullable */
  documentName?: string | null;
  status: UserSignRequestStatus;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface SignRequestInput {
  otpCode: string;
  /** Passphrase Digital ID pengguna */
  passphrase: string;
  /** Page number (1-indexed) */
  signaturePage: number;
  /** X position as percentage (0-100) */
  signatureX: number;
  /** Y position as percentage (0-100) */
  signatureY: number;
  /** Width as percentage of page width */
  signatureWidth: number;
  /** Height as percentage of page height */
  signatureHeight: number;
}

export type SignRequestStatus = typeof SignRequestStatus[keyof typeof SignRequestStatus];


export const SignRequestStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface SignRequest {
  id: number;
  documentId: number;
  userId: number;
  userName?: string;
  documentName?: string;
  status: SignRequestStatus;
  signaturePage: number;
  signatureX: number;
  signatureY: number;
  signatureWidth?: number;
  signatureHeight?: number;
  /** @nullable */
  rejectionReason?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface ActivityItem {
  id: number;
  type: string;
  description: string;
  createdAt: string;
}

export interface AdminDashboard {
  totalUsers: number;
  pendingDigitalIdRequests: number;
  pendingSignRequests: number;
  totalDocuments: number;
  approvedDigitalIds: number;
  totalSignedDocuments: number;
  recentActivity?: ActivityItem[];
}

export interface RejectInput {
  reason: string;
}

export interface VerifyInput {
  /** Base64-encoded document to verify */
  fileData: string;
}

export interface VerifyResult {
  isValid: boolean;
  message: string;
  /** @nullable */
  signerName?: string | null;
  /** @nullable */
  signerEmail?: string | null;
  /** @nullable */
  signedAt?: string | null;
  details?: string;
}

export interface VerifyQrResult {
  isValid: boolean;
  documentName: string;
  signerName: string;
  signerEmail: string;
  signerRole?: string;
  /** @nullable */
  nimNip?: string | null;
  signedAt: string;
}