# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Unila Digital Sign API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
  /auth/change-password:
    post:
      operationId: changePassword
      tags: [auth]
      summary: Change password (requires old password + OTP)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Invalid old password or OTP
        "401":
          description: Unauthorized

  # OTP
...
  /digital-id/request:
    post:
      operationId: requestDigitalId
      tags: [digital-id]
      summary: Request a new digital ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DigitalIdInput"
      responses:
        "201":
          description: Digital ID request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalIdRequest"
        "400":
          description: Invalid input or OTP not setup
  /digital-id/status:
    get:
      operationId: getDigitalIdStatus
      tags: [digital-id]
      summary: Get current user digital ID request status
      responses:
        "200":
          description: Digital ID status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DigitalIdRequest"
        "404":
          description: No request found
  /digital-id/download:
    get:
      operationId: downloadDigitalId
      tags: [digital-id]
      summary: Download approved digital ID (.p12 file)
      responses:
        "200":
          description: Digital ID file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          description: Not approved yet

  # Signature image
...
  /documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List current user documents
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"

    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload a document for signing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
  /documents/{id}:
    get:
      operationId: getDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get document details
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "404":
          description: Not found

    delete:
      operationId: deleteDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Delete a document
      responses:
        "200":
          description: Deleted
  /sign-requests:
    get:
      operationId: listSignRequests
      tags: [documents]
      summary: List current user's own sign requests
      responses:
        "200":
          description: User sign requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserSignRequest"
        "401":
          description: Unauthorized
  /documents/{id}/sign:
    post:
      operationId: signDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Request document signing with OTP verification and signature position
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignRequestInput"
      responses:
        "201":
          description: Sign request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignRequest"
        "400":
          description: Invalid OTP or missing digital ID
  /documents/{id}/download:
    get:
      operationId: downloadSignedDocument
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Download signed document (if approved)
      responses:
        "200":
          description: Signed document file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          description: Not signed yet
  /documents/{id}/sign-status:
    get:
      operationId: getSignStatus
      tags: [documents]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      summary: Get sign request status for a document
      responses:
        "200":
          description: Sign request status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignRequest"
        "404":
          description: No sign request

  # Admin
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        otpRequired:
          type: boolean
    User:
      type: object
      required: [id, email, name, role, otpEnabled]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, dosen, mahasiswa]
        otpEnabled:
          type: boolean
        digitalIdStatus:
          type: ["string", "null"]
          enum: [pending, approved, rejected, revoked, "null"]
        hasSignature:
          type: boolean
        createdAt:
          type: string
          format: date-time

    # OTP
...
    ChangePasswordInput:
      type: object
      required: [oldPassword, newPassword, otpCode]
      properties:
        oldPassword:
          type: string
        newPassword:
          type: string
        otpCode:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    DigitalIdInput:
      type: object
      required: [name, role, email, passphrase]
      properties:
        name:
          type: string
        role:
          type: string
          enum: [dosen, mahasiswa]
        email:
          type: string
          format: email
        passphrase:
          type: string
        nimNip:
          type: string
          description: NIM (mahasiswa) or NIP (dosen)
    DigitalIdRequest:
      type: object
      required: [id, userId, name, role, email, status, isApproved, isReady, isSent, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [dosen, mahasiswa]
        email:
          type: string
        nimNip:
          type: ["string", "null"]
          description: NIM (mahasiswa) or NIP (dosen)
        status:
          type: string
          enum: [pending, approved, rejected, revoked]
        isApproved:
          type: boolean
        isReady:
          type: boolean
        isSent:
          type: boolean
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Signature
...
    DocumentInput:
      type: object
      required: [name, fileData, fileType]
      properties:
        name:
          type: string
        fileData:
          type: string
          description: Base64-encoded file
        fileType:
          type: string
        pageCount:
          type: integer
    Document:
      type: object
      required: [id, userId, name, fileType, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        fileType:
          type: string
        status:
          type: string
          enum: [uploaded, pending_sign, signed, rejected]
        pageCount:
          type: ["integer", "null"]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserSignRequest:
      type: object
      required: [id, documentId, status, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        documentName:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SignRequestInput:
      type: object
      required: [otpCode, passphrase, signaturePage, signatureX, signatureY, signatureWidth, signatureHeight]
      properties:
        otpCode:
          type: string
        passphrase:
          type: string
          description: Passphrase Digital ID pengguna
        signaturePage:
          type: integer
          description: Page number (1-indexed)
        signatureX:
          type: number
          description: X position as percentage (0-100)
        signatureY:
          type: number
          description: Y position as percentage (0-100)
        signatureWidth:
          type: number
          description: Width as percentage of page width
        signatureHeight:
          type: number
          description: Height as percentage of page height
    SignRequest:
      type: object
      required: [id, documentId, userId, status, signaturePage, signatureX, signatureY, createdAt]
      properties:
        id:
          type: integer
        documentId:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        documentName:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        signaturePage:
          type: integer
        signatureX:
          type: number
        signatureY:
          type: number
        signatureWidth:
          type: number
        signatureHeight:
          type: number
        rejectionReason:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # Admin
...